added a SizeLimit configuration option and bumped other "limit" options to Int64

+++ hedgehog/src/Hedgehog/Internal/Range.hs
@@ -40,6 +40,7 @@ module Hedgehog.Internal.Range (
   , scaleExponentialFloat
   ) where
 
+import           Data.Int (Int64)
 import           Data.Bifunctor (bimap)
 
 import           Prelude hiding (minimum, maximum)
@@ -53,7 +54,7 @@ import           Prelude hiding (minimum, maximum)
 --
 newtype Size =
   Size {
-      unSize :: Int
+      unSize :: Int64
     } deriving (Eq, Ord, Num, Real, Enum, Integral)
 
 instance Show Size where
